refactor: Move git related command options to new namespace to follow convention
