+++ Corgibytes.Freshli.Cli/CommandRunners/GitCommandRunner.cs
@@ -1,6 +1,7 @@
 using System;
 using System.CommandLine.Invocation;
 using Corgibytes.Freshli.Cli.CommandOptions;
+using Corgibytes.Freshli.Cli.CommandOptions.Git;
 using Corgibytes.Freshli.Cli.Commands;
 using Corgibytes.Freshli.Lib;
