+++ Corgibytes.Freshli.Cli/Commands/Git/GitCloneCommand.cs
@@ -1,5 +1,6 @@
 using System.CommandLine;
 using Corgibytes.Freshli.Cli.CommandOptions;
+using Corgibytes.Freshli.Cli.CommandOptions.Git;
 
 namespace Corgibytes.Freshli.Cli.Commands.Git;
